changed data type of id.

+++ springboot-backend/src/main/java/com/labapp/repository/LaborantRepository.java
@@ -4,8 +4,9 @@ import com.labapp.entity.Laborant;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface LaborantRepository extends JpaRepository<Laborant, Integer> {
+public interface LaborantRepository extends JpaRepository<Laborant, Long> {
 
     @Query("SELECT lab FROM Laborant lab WHERE lab.hospitalIdNo = ?1")
     Laborant findLaborantByHospitalIdNo(String hospitalIdNo);
+
 }
