+++ springboot-backend/src/main/java/com/labapp/service/LaborantService.java
@@ -1,19 +1,21 @@
 package com.labapp.service;
 
 import com.labapp.entity.Laborant;
+import org.springframework.security.core.userdetails.UserDetailsService;
 
 import java.util.List;
 
-public interface LaborantService {
+public interface LaborantService{
 
     List<Laborant> findAll();
 
-    Laborant findById(Integer id);
+    Laborant findById(Long id);
 
     Laborant save(Laborant theLaborant);
 
-    void deleteById(Integer id);
+    void deleteById(Long id);
 
     Laborant findLaborantByHospitalIdNo(String hospitalIdNo);
+
 
 }
